+++ server/src/test/java/net/oneandone/stool/kubernetes/EngineIT.java
@@ -51,21 +51,6 @@ public class EngineIT {
     //-- pods
 
     @Test
-    public void helm() throws IOException {
-        V1Secret s;
-        byte[] release;
-
-
-        try (Engine engine = Engine.create(WORLD, "local", Strings.toMap())) {
-            for (String name : engine.secretList().keySet()) {
-                if (name.startsWith("sh.helm.release.v1.")) {
-                    System.out.println("name: " + name);
-                    System.out.println("values: " + engine.helmRead(name).get("chart").getAsJsonObject().get("values"));
-                }
-            }
-        }
-    }
-    @Test
     public void podTerminating() throws IOException {
         final String name = "pod";
         Collection<PodInfo> lst;